requirements.txt
openai>=1.0.0
alpaca-trade-api>=3.0.0
alpaca-py>=0.8.0
python-dotenv>=1.0.0
pytz>=2023.3
aiohttp>=3.9.0
